+++ cli/src/merge_tools/diff_working_copies.rs
@@ -16,6 +16,7 @@ use jj_lib::local_working_copy::TreeStateError;
 use jj_lib::local_working_copy::TreeStateSettings;
 use jj_lib::matchers::EverythingMatcher;
 use jj_lib::matchers::Matcher;
+use jj_lib::matchers::NothingMatcher;
 use jj_lib::merge::Diff;
 use jj_lib::merged_tree::MergedTree;
 use jj_lib::merged_tree::TreeDiffEntry;
@@ -283,6 +284,7 @@ diff editing in mind and be a little inaccurate.
             base_ignores,
             progress: None,
             start_tracking_matcher: &EverythingMatcher,
+            force_tracking_matcher: &NothingMatcher,
             max_new_file_size: u64::MAX,
         })?;
         Ok(output_tree_state.current_tree().clone())